Re-arranged files to extract homepage welcome and delete unused component
