+++ pages/index.tsx
@@ -1,18 +1,21 @@
-import { Box, useMediaQuery, useTheme } from '@mui/material';
+import { Box, Typography, useMediaQuery, useTheme } from '@mui/material';
+import type { Theme } from '@mui/system';
+import { useRouter } from 'next/router';
+import type { ReactElement } from 'react';
+import CatCardFlipper from '../components/catCardFlipper/CatCardFlipper';
+import FlowerFooter from '../components/flowerFooter/FlowerFooter';
 import { Header } from '../components/header';
 import { LoadingPage } from '../components/loadingPage/LoadingPage';
 import { MagicCatReload } from '../components/magicCatAnimation/magicCatReload/MagicCatReload';
+import RaisedButton from '../components/raisedButton/RaisedButton';
 import { TarotDeckContext } from '../context/TarotDeckContext';
-import { useTarotCard } from '../hooks/useTarotCard';
-import FlowerFooter from '../components/flowerFooter/FlowerFooter';
-import HomepageWelcome from '../components/homepageWelcome/HomepageWelcome';
 import useFetchTarotDeck from '../hooks/fetchTarotDeck';
-import type { Theme } from '@mui/system';
-import type { ReactElement } from 'react';
+import { useTarotCard } from '../hooks/useTarotCard';
 
 const homePageMainImage = '/image/homepageImage.webp';
 
 export default function Home(): ReactElement {
+  const router = useRouter();
   const { tarotDeck, loading } = useFetchTarotDeck();
   const { tarotCardData } = useTarotCard(tarotDeck);
   const { palette } = useTheme();
@@ -55,7 +58,54 @@ export default function Home(): ReactElement {
               }}
             />
           </Box>
-          <HomepageWelcome />
+          <Box
+            sx={{
+              p: 4,
+              borderRadius: 3
+            }}
+          >
+            <Box
+              sx={{
+                display: 'flex',
+                justifyContent: 'alignItems',
+                alignContent: 'center',
+                alignItems: 'center',
+                flexDirection: isSmallScreen ? 'column' : 'row'
+              }}
+            >
+              <CatCardFlipper />
+              <Box
+                sx={{
+                  display: 'flex',
+                  flexDirection: 'column',
+                  gap: 4,
+                  textAlign: 'center',
+                  justifyContent: 'center',
+                  alignItems: 'center'
+                }}
+              >
+                <Typography variant='h6'>
+                  Considerate Cat is a tarot deck that was published in 2018,
+                  brought to life by a community of folks on Kickstarter. Since
+                  then, the deck has been brought into many inspiring
+                  people&apos;s homes, featuring stories of rescue cats from
+                  across the world. Now, we are bringing the deck to life
+                  online! Here, you can have your cards read, learn about the
+                  cats that inspired the deck, and dive deeper into the artwork
+                </Typography>
+                <Typography variant='h6'>
+                  Also, Butters is patiently waiting to have his cards read.
+                  Click him to give him the reading he&apos;s been waiting for!
+                </Typography>
+                <RaisedButton
+                  buttonLabel='Get My Own Tarot Read!'
+                  onClick={() => {
+                    router.push('/TarotReadPage');
+                  }}
+                />
+              </Box>
+            </Box>
+          </Box>
         </Box>
         <FlowerFooter />
       </Box>
